// GamerHub_Store/src/app/modules/management/models/eventopar.ts
export class Eventopar {
    nameevent: string;
    acargo: string;
    telefono: string;
    email: string;
    date: string;
    empieza: string;
    termina: string;
    city: string;
    horario: string;

    constructor (nameevent: string, acargo: string, telefono: string, email: string, date: string, city: string, empieza: string, termina: string, horario: string) {
    this.nameevent = '';
    this.acargo = '';
    this.telefono = '';
    this.email = '';
    this.date = '';
    this.city = '';
    this.empieza = '';
    this.termina = '';
    this.horario = '';
    }
}